show-installer: support wget

+++ cmd/norouter/show_installer.go
@@ -38,16 +38,55 @@ var showInstallerCommand = &cli.Command{
 
 func showInstallerAction(clicontext *cli.Context) error {
 	tmpl := `#!/bin/sh
-set -eux
+set -eu
 # Installation script for NoRouter
 # NOTE: make sure to use the same version across all the hosts.
 version="{{.Version}}"
+echo "# Version: ${version}"
+
 bindir="$HOME/bin"
+echo "# Destination: ${bindir}/norouter"
 mkdir -p "${bindir}"
-rm -f "${bindir}/norouter"
-curl -o "${bindir}/norouter" --fail -L https://github.com/norouter/norouter/releases/download/v${version}/norouter-$(uname -s)-$(uname -m)
-chmod +x "${bindir}/norouter"
-echo "Successfully installed ${bindir}/norouter (version ${version})"
+
+download(){
+  local="$1"
+	remote="$2"
+	echo "# Downloading ${remote}"
+	if command -v curl >/dev/null 2>&1; then
+		( set -x; curl -fL -o "$local" "$remote" )
+	elif command -v wget >/dev/null 2>&1; then
+	  ( set -x; wget -O "$local" "$remote" )
+  else
+	  echo >&2 "curl or wget needs to be installed"
+		exit 1
+	fi
+}
+
+fname="norouter-$(uname -s)-$(uname -m)"
+tmp=$(mktemp -d)
+download "${tmp}/${fname}" "https://github.com/norouter/norouter/releases/download/v${version}/${fname}"
+chmod +x "${tmp}/${fname}"
+download "${tmp}/SHA256SUMS" "https://github.com/norouter/norouter/releases/download/v${version}/SHA256SUMS"
+
+if command -v sha256sum &> /dev/null; then
+  (
+	  cd "${tmp}"
+		echo "# Printing sha256sum of SHA256SUMS file itself"
+		sha256sum SHA256SUMS
+		echo "# Checking SHA256SUMS"
+    grep "${fname}" SHA256SUMS | sha256sum -c -
+	)
+fi
+if [ -x "${bindir}/norouter" ]; then
+  echo "# Removing existing ${bindir}/norouter"
+  rm -f "${bindir}/norouter"
+fi
+echo "# Installing ${fname} onto ${bindir}/norouter"
+mv "${tmp}/${fname}" "${bindir}/norouter"
+
+rm -rf "${tmp}"
+
+echo "# Successfully installed ${bindir}/norouter (version ${version})"
 `
 	m := map[string]string{
 		"Version": clicontext.String("version"),
